refactor(δ): move Practice/ChatMessage canonical types into wire modules
Inverted the dependency: wire modules now own the type declarations,
worker/types.ts re-exports them. Previously the wire modules just
re-exported from worker/types.ts.

Moved into worker/types/wire/:
- practice.ts:    Practice, Workspace, PracticeOrWorkspace,
                  ConversationConfig, PracticeConfig,
                  SubscriptionLifecycleStatus
- conversation.ts: ChatMessage, ChatSession, MessageReaction,
                  BackendConversation

worker/types.ts:
- `import type` then `export type` so this file can still extend
  ChatMessage in ChatMessageUI (`extends` needs the name in local
  scope, not just re-exported)
- ChatMessageUI stays here because UIMessageExtras is worker-internal
  scaffolding, not wire shape

Frontend consumers continue importing from @/shared/types/wire
unchanged. Worker code that did `import type { Practice } from '../types'`
keeps working via the re-export bridge.

Part of phase δ — RemoteApiService gateway typing.
